+++ .vitepress/theme/components/BlogArticleLayout.module.css
@@ -91,6 +91,101 @@
   display: none;
 }
 
+.content :global(> *) {
+  max-width: 74ch;
+}
+
+.content :global(> h2),
+.content :global(> h3),
+.content :global(> h4),
+.content :global(> h5),
+.content :global(> h6),
+.content :global(> pre),
+.content :global(> div[class*="language-"]),
+.content :global(> blockquote),
+.content :global(> table),
+.content :global(> .mermaid) {
+  max-width: 100%;
+}
+
+.content :global(p),
+.content :global(ul),
+.content :global(ol),
+.content :global(blockquote) {
+  color: var(--vp-c-text-2);
+  font-size: 1.06rem;
+  line-height: 1.9;
+}
+
+.content :global(p) {
+  margin: 1rem 0;
+  text-wrap: pretty;
+}
+
+.content :global(ul),
+.content :global(ol) {
+  margin: 1.1rem 0;
+  padding-left: 1.45rem;
+}
+
+.content :global(li) {
+  margin: 0.45rem 0;
+}
+
+.content :global(li > p) {
+  margin: 0.45rem 0;
+}
+
+.content :global(h2),
+.content :global(h3),
+.content :global(h4) {
+  color: #fff;
+  font-family: "Lato", sans-serif;
+  line-height: 1.15;
+  text-wrap: balance;
+}
+
+.content :global(h2) {
+  margin-top: 2.75rem;
+  margin-bottom: 1rem;
+  font-size: clamp(1.7rem, 2.4vw, 2.3rem);
+}
+
+.content :global(h3) {
+  margin-top: 2.1rem;
+  margin-bottom: 0.85rem;
+  font-size: clamp(1.35rem, 1.8vw, 1.7rem);
+}
+
+.content :global(h4) {
+  margin-top: 1.7rem;
+  margin-bottom: 0.7rem;
+  font-size: 1.08rem;
+  letter-spacing: 0.01em;
+}
+
+.content :global(blockquote) {
+  margin: 1.5rem 0;
+}
+
+.content :global(code) {
+  font-size: 0.95em;
+}
+
+.content :global(hr) {
+  max-width: 74ch;
+  margin: 2rem 0;
+}
+
+.content :global(img) {
+  border-radius: 1rem;
+  margin: 1.75rem 0;
+}
+
+.content :global(table) {
+  margin: 1.5rem 0;
+}
+
 .aside {
   position: sticky;
   top: calc(var(--vp-nav-height) + 2rem);
@@ -144,6 +239,10 @@
 @media (max-width: 1279px) {
   .body {
     grid-template-columns: minmax(0, 1fr);
+  }
+
+  .content :global(> *) {
+    max-width: 100%;
   }
 
   .aside {
